requirements.txt
openai
python-dotenv
PyYAML
pandas
matplotlib